git-svn-id: http://miranda.googlecode.com/svn/branches/stable@8281 83df23c2-51fd-11de-bb24-2d5e9b4af41d

+++ miranda/src/modules/langpack/langpack.c
@@ -75,7 +75,7 @@ static int IsEmpty(char *str)
 static void ConvertBackslashes(char *str)
 {
 	char *pstr;
-	for(pstr=str;*pstr;pstr=CharNextExA(langPack.defaultANSICp,pstr,0)) {
+	for(pstr=str;*pstr;pstr=CharNextExA((WORD)langPack.defaultANSICp,pstr,0)) {
 		if(*pstr=='\\') {
 			switch(pstr[1]) {
 case 'n': *pstr='\n'; break;
